requirements.gpu.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
redis==5.0.1
celery==5.3.6
# GPU-enabled PyTorch (CUDA 11.8)
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1+cu118
torchvision==0.18.1+cu118
torchaudio==2.3.1+cu118
transformers>=4.52.0
diffusers>=0.36.0
accelerate>=0.30.0
huggingface-hub>=0.20.0
pillow==10.2.0
numpy==1.25.2
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
slowapi==0.1.9
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
boto3==1.34.34
python-dotenv==1.0.0
httpx==0.26.0
mcp

# Chatterbox (installed separately with --no-deps to avoid torch version conflicts)
librosa==0.11.0
s3tokenizer
resemble-perth==1.0.1
conformer==0.3.2
safetensors==0.5.3
omegaconf
pyloudnorm

